requirements.txt
# Core dependencies
pydantic>=2.10
typer>=0.15
pyyaml>=6.0
pandas>=2.0
aiohttp>=3.9
deepdiff>=6.0
langfuse>=2.57
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Development dependencies (install with: pip install -r requirements-dev.txt)
# pytest>=8.3
# ruff>=0.8
# mypy>=1.8

# Optional LLM dependencies (install with: pip install -r requirements-llm.txt)
# openai>=1.59
# anthropic>=0.18
